+++ cmd/waarp-gateway/url.go
@@ -7,7 +7,8 @@ import (
 )
 
 type addrOpt struct {
-	Address gwAddr `short:"a" long:"address" description:"The address of the gateway" env:"WAARP_GATEWAY_ADDRESS"`
+	Address  gwAddr `short:"a" long:"address" description:"The address of the gateway" env:"WAARP_GATEWAY_ADDRESS"`
+	Insecure bool   `short:"i" long:"insecure" description:"Skip certificate verification" env:"WAARP_GATEWAY_INSECURE"`
 }
 
 type gwAddr struct{}