+++ doc/source/changelog.rst
@@ -3,6 +3,8 @@
 Historique des versions
 =======================
 
+* :bug:`248` Ajout de l'option `insecure` au client terminal afin de désactiver la 
+  vérification des certificats serveur https.
 * :release:`0.3.1 <2021-01-25>`
 * :bug:`241` Correction du typage de la colonne `permissions` de la table `users`.
   La colonne est désormais de type *BINARY* (au lieu de *INT*).